// apps/web/generated/api/models/budgetType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */

export type BudgetType = typeof BudgetType[keyof typeof BudgetType];


export const BudgetType = {
  fixed: 'fixed',
  range: 'range',
  hourly: 'hourly',
  negotiable: 'negotiable',
} as const;