// apps/web/generated/api/models/deliveryStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */

export type DeliveryStatus = typeof DeliveryStatus[keyof typeof DeliveryStatus];


export const DeliveryStatus = {
  submitted: 'submitted',
  under_review: 'under_review',
  approved: 'approved',
  rejected: 'rejected',
  revised: 'revised',
  superseded: 'superseded',
} as const;