// apps/web/generated/api/models/projectApplicationStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */

export type ProjectApplicationStatus = typeof ProjectApplicationStatus[keyof typeof ProjectApplicationStatus];


export const ProjectApplicationStatus = {
  applied: 'applied',
  shortlisted: 'shortlisted',
  accepted: 'accepted',
  rejected: 'rejected',
  withdrawn: 'withdrawn',
  expired: 'expired',
} as const;