// apps/web/generated/api/models/reviewStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */

export type ReviewStatus = typeof ReviewStatus[keyof typeof ReviewStatus];


export const ReviewStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;